+++ pypgx/sdk/__init__.py
@@ -1,3 +1,3 @@
-from .utils import (Archive, add_cn_samples, compare_metadata, simulate_copy_number)
+from .utils import (Archive, add_cn_samples, compare_metadata, get_bundle_path, simulate_copy_number)
 
-__all__ = ['Archive', 'add_cn_samples', 'compare_metadata', 'simulate_copy_number']
+__all__ = ['Archive', 'add_cn_samples', 'compare_metadata', 'get_bundle_path', 'simulate_copy_number']